Update js/bootstrap-tab.js
I came across this issue this morning when integrating bootstrap with a script that was using ajax. 

Inside a tab is a form with a submit button. Bootstrap tab is disabling the submit button action.

Commenting //e.preventDefault() solves the issue but I am not sure if this is the correct way of fixing it.

Thank you